DRY up error_wrapper.go
